Changed datepickers and Calendar modelValue type.
It's possible to use String and Date as model value of datepicker and calendar components now.
To use string as model value, you have to provide dateFormat props which sets format of the returned string value.

Min and Max dates should be same type as modelValue. Different types will not break component, but may lead to unexpected behavior.